update: changed location of object log from core to utils/log.py
